requirements.txt
requests
pandas
psycopg2-binary
boto3
kafka-python==2.3.1